@@ -1,74 +1,71 @@
 {
 	"name": "@marianmeres/stuic",
-	"version": "1.126.0",
+	"version": "2.0.1",
 	"scripts": {
 		"dev": "vite dev",
-		"build": "vite build && npm run package && node ./scripts/date.js",
-		"build:watch": "fswatch -o src | xargs -n1 -I{} npm run build",
+		"build": "vite build && npm run prepack",
 		"preview": "vite preview",
-		"package": "svelte-kit sync && svelte-package && publint",
-		"prepublishOnly": "npm run package",
+		"prepare": "svelte-kit sync || echo ''",
+		"prepack": "svelte-kit sync && svelte-package && publint",
+		"package": "npm run prepack",
+		"package:watch": "svelte-kit sync && svelte-package --watch && publint",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"lint": "prettier --check .",
 		"format": "prettier --write .",
-		"prettier": "npm run format",
-		"release:patch": "release -v patch",
-		"release": "release -v minor",
-		"test": "vitest"
-	},
-	"exports": {
-		".": {
-			"types": "./dist/index.d.ts",
-			"svelte": "./dist/index.js"
-		}
+		"lint": "prettier --check .",
+		"test": "vitest --dir src/"
 	},
 	"files": [
 		"dist",
 		"!dist/**/*.test.*",
 		"!dist/**/*.spec.*"
 	],
-	"peerDependencies": {
-		"svelte": "^4.0.0 || ^5.0.0-next.1"
-	},
-	"devDependencies": {
-		"@marianmeres/parse-boolean": "^1.1.7",
-		"@marianmeres/random-human-readable": "^1.6.1",
-		"@marianmeres/release": "^1.1.2",
-		"@sveltejs/adapter-auto": "^2.0.0",
-		"@sveltejs/kit": "^1.27.4",
-		"@sveltejs/package": "^2.0.0",
-		"@tailwindcss/forms": "^0.5.7",
-		"@types/node": "^20.11.0",
-		"autoprefixer": "^10.4.16",
-		"clsx": "^2.0.0",
-		"nodemon": "^3.0.2",
-		"postcss": "^8.4.32",
-		"prettier": "^3.0.0",
-		"prettier-plugin-svelte": "^3.0.0",
-		"publint": "^0.1.9",
-		"sass": "^1.69.5",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"tailwindcss": "^3.3.6",
-		"ts-node": "^10.9.2",
-		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^4.4.2",
-		"vitest": "^1.1.3"
-	},
+	"sideEffects": [
+		"**/*.css"
+	],
 	"svelte": "./dist/index.js",
 	"types": "./dist/index.d.ts",
 	"type": "module",
+	"exports": {
+		".": {
+			"types": "./dist/index.d.ts",
+			"svelte": "./dist/index.js"
+		}
+	},
+	"peerDependencies": {
+		"svelte": "^5.0.0"
+	},
+	"devDependencies": {
+		"@marianmeres/icons-fns": "^4.3.1",
+		"@marianmeres/random-human-readable": "^1.6.1",
+		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/kit": "^2.36.1",
+		"@sveltejs/package": "^2.5.0",
+		"@sveltejs/vite-plugin-svelte": "^5.1.1",
+		"@tailwindcss/cli": "^4.1.12",
+		"@tailwindcss/forms": "^0.5.10",
+		"@tailwindcss/typography": "^0.5.16",
+		"@tailwindcss/vite": "^4.1.12",
+		"dotenv": "^16.6.1",
+		"prettier": "^3.6.2",
+		"prettier-plugin-svelte": "^3.4.0",
+		"publint": "^0.3.12",
+		"svelte": "^5.38.2",
+		"svelte-check": "^4.3.1",
+		"tailwindcss": "^4.1.12",
+		"typescript": "^5.9.2",
+		"vite": "^6.3.5",
+		"vitest": "^3.2.4"
+	},
+	"packageManager": "pnpm@10.4.1+sha512.c753b6c3ad7afa13af388fa6d808035a008e30ea9993f58c6663e2bc5ff21679aa834db094987129aa4d488b86df57f7b634981b2f827cdcacc698cc0cfb88af",
 	"dependencies": {
-		"@marianmeres/clog": "^1.0.1",
-		"@marianmeres/icons-fns": "^4.0.4",
-		"@marianmeres/store": "^1.5.0",
-		"@marianmeres/switch-store": "^1.3.1",
-		"@marianmeres/ticker": "^1.5.0",
-		"dset": "^3.1.3",
-		"esm-env": "^1.0.0",
-		"klona": "^2.0.6",
-		"tailwind-merge": "^2.1.0"
+		"@marianmeres/clog": "^2.3.3",
+		"@marianmeres/item-collection": "^1.2.19",
+		"@marianmeres/parse-boolean": "^1.1.7",
+		"@marianmeres/ticker": "^1.15.0",
+		"clsx": "^2.1.1",
+		"esm-env": "^1.2.2",
+		"runed": "^0.23.4",
+		"tailwind-merge": "^3.3.1"
 	}
 }
