@@ -1,37 +1,29 @@
-<script context="module" lang="ts">
-	import { twMerge2 } from '../../utils/tw-merge2.js';
-	import Thc from '../Thc/Thc.svelte';
-
-	export interface FieldsetConfigClasses {
-		box?: string;
-		legend?: string;
-	}
-
-	const _PRESET: FieldsetConfigClasses = {
-		box: `border border-neutral-200 p-4 pt-3 rounded-md`,
-		legend: `px-2`,
-	};
-
-	export class FieldsetConfig {
-		static class: FieldsetConfigClasses = {};
-	}
-</script>
-
 <script lang="ts">
-	let _class: FieldsetConfigClasses = {};
-	export { _class as class };
+	import type { Snippet } from "svelte";
+	import type { THC } from "../Thc/Thc.svelte";
+	import Thc from "../Thc/Thc.svelte";
+	import { twMerge } from "../../utils/tw-merge.js";
 
-	export let legend = '';
-
-	const _collectClasses = (k: keyof FieldsetConfigClasses, extra = '') =>
-		[_PRESET?.[k] || '', FieldsetConfig?.class?.[k] || '', _class?.[k] || ''].join(' ');
+	interface Props {
+		legend?: THC;
+		class?: string;
+		classLegend?: string;
+		children?: Snippet;
+	}
+	let { legend, class: classProp, children, classLegend }: Props = $props();
 </script>
 
-<fieldset class={twMerge2(_collectClasses('box'))}>
+<fieldset
+	class={twMerge(
+		"stuic-fieldset",
+		"border border-neutral-300 p-4 pt-3 rounded-md my-8",
+		classProp
+	)}
+>
 	{#if legend}
-		<legend class={twMerge2(_collectClasses('legend'))}>
+		<legend class={twMerge("px-2", classLegend)}>
 			<Thc thc={legend} forceAsHtml />
 		</legend>
 	{/if}
-	<slot />
+	{@render children?.()}
 </fieldset>
