@@ -1,12 +1,14 @@
 <script lang="ts">
-	import { twMerge2 } from '../../utils/tw-merge2.js';
+	import { twMerge } from "$lib/index.js";
 
-	let _class = '';
-	export { _class as class };
+	interface Props {
+		class?: string;
+		strokeWidth?: 0.5 | 1 | 1.5 | 2 | 2.5 | 3 | 3.5 | 4;
+	}
 
-	export let strokeWidth: 0.5 | 1 | 1.5 | 2 | 2.5 | 3 | 3.5 | 4 = 2;
+	let { class: classProps, strokeWidth = 2 }: Props = $props();
 
-	// size-6 is 1.5rem, 24px
+	// size-6 = 1.5rem = 24px
 </script>
 
 <svg
@@ -14,7 +16,7 @@
 	viewBox="0 0 24 24"
 	stroke-width={strokeWidth}
 	stroke="currentColor"
-	class={twMerge2(`inline size-6 ${_class}`)}
+	class={twMerge("inline size-6", classProps)}
 >
 	<path stroke-linecap="round" stroke-linejoin="round" d="M6 18 18 6M6 6l12 12" />
 </svg>
