@@ -1,14 +1,13 @@
-import { iconFeatherAlertTriangle } from '@marianmeres/icons-fns/feather/iconFeatherAlertTriangle.js';
-import { iconFeatherCheckCircle } from '@marianmeres/icons-fns/feather/iconFeatherCheckCircle.js';
-import { iconFeatherInfo } from '@marianmeres/icons-fns/feather/iconFeatherInfo.js';
-import { iconFeatherRefreshCw } from '@marianmeres/icons-fns/feather/iconFeatherRefreshCw.js';
-import { iconFeatherXOctagon } from '@marianmeres/icons-fns/feather/iconFeatherXOctagon.js';
-import type { AlertConfirmPromptIcons } from './AlertConfirmPrompt.svelte';
+import { iconFeatherAlertTriangle } from "@marianmeres/icons-fns/feather/iconFeatherAlertTriangle.js";
+import { iconFeatherCheckCircle } from "@marianmeres/icons-fns/feather/iconFeatherCheckCircle.js";
+import { iconFeatherInfo } from "@marianmeres/icons-fns/feather/iconFeatherInfo.js";
+import { iconFeatherRefreshCw } from "@marianmeres/icons-fns/feather/iconFeatherRefreshCw.js";
+import { iconFeatherXOctagon } from "@marianmeres/icons-fns/feather/iconFeatherXOctagon.js";
 
-export const acpDefaultIcons: AlertConfirmPromptIcons = {
-	info: () => iconFeatherInfo({}),
+export const acpDefaultIcons = {
+	info: () => iconFeatherInfo({ size: 24 }),
 	success: () => iconFeatherCheckCircle({}),
-	warn: () => iconFeatherAlertTriangle({ class: '-mt-[3px]' }), // move up a little because it looks better with the triangle
+	warn: () => iconFeatherAlertTriangle({ class: "-mt-[3px]" }), // move up a little because it looks better with the triangle
 	error: () => iconFeatherXOctagon({}),
-	spinner: () => iconFeatherRefreshCw({ size: 32, class: 'opacity-50' }),
+	spinner: () => iconFeatherRefreshCw({ size: 32, class: "opacity-50" }),
 };
