@@ -1,3 +1,5 @@
 let _id = 0;
 
-export const getId = (prefix: string = 'id-') => `${prefix}${++_id}`;
+export function getId(prefix: string = "id-") {
+	return `${prefix}${++_id}`;
+}
