@@ -1,13 +1,14 @@
-import { sveltekit } from '@sveltejs/kit/vite';
+import { sveltekit } from "@sveltejs/kit/vite";
 // import { defineConfig } from 'vite';
-import { defineConfig } from 'vitest/config';
+import { defineConfig } from "vitest/config";
+import tailwindcss from "@tailwindcss/vite";
+import dotenv from "dotenv";
+
+dotenv.config();
 
 export default defineConfig({
-	plugins: [sveltekit()],
+	plugins: [tailwindcss(), sveltekit()],
 	server: {
-		port: 5180,
-	},
-	test: {
-		include: ['src/**/*.{test,spec}.{js,ts}'],
+		port: parseInt(process.env.DEV_PORT || "8888"),
 	},
 });
